Reuse activation maps from public s3 for seed_test_run
Speeds up `seed_test_run` by using pregenerated activation maps.
Moves the activation map generation and uploading to a seperate
folder.
